(refactor) new .check package in model.slam, as for model.pisdf and model.scenario
