+++ plugins/org.preesm.algorithm/src/org/preesm/algorithm/pisdf/autodelays/AutoDelaysTask.java
@@ -92,7 +92,7 @@ import org.preesm.model.scenario.ScenarioConstants;
 import org.preesm.model.scenario.SimulationInfo;
 import org.preesm.model.slam.Component;
 import org.preesm.model.slam.Design;
-import org.preesm.model.slam.utils.SlamDesignPEtypeChecker;
+import org.preesm.model.slam.check.SlamDesignPEtypeChecker;
 import org.preesm.workflow.elements.Workflow;
 import org.preesm.workflow.implement.AbstractTaskImplementation;
 import org.preesm.workflow.implement.AbstractWorkflowNodeImplementation;
