+++ plugins/org.preesm.algorithm/src/org/preesm/algorithm/synthesis/memalloc/LegacyMemoryAllocation.java
@@ -79,7 +79,7 @@ import org.preesm.model.pisdf.PiGraph;
 import org.preesm.model.scenario.Scenario;
 import org.preesm.model.slam.ComponentInstance;
 import org.preesm.model.slam.Design;
-import org.preesm.model.slam.utils.SlamDesignPEtypeChecker;
+import org.preesm.model.slam.check.SlamDesignPEtypeChecker;
 
 /**
  *
